tests/wpt/tests/html/dom/render-blocking/element-render-blocking-011.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Rel attribute changes in the head to apply</title>

<link id=link rel=stylesheet href="#last" blocking="render">
<script>
link.rel = "expect";

async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_true(!!document.getElementById("last")));
    t.done();
  });
}, "changing rel to expect in the head causes it to be blocking");
</script>
</head>
<body>
  <div id="first"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="second"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="last"></div>
</body>
